Move zpages to package zpages instead of internal
