+++ zpages/zpagesserver.go
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package zpagesserver
+// Package zpages contains the config on setting up ZPages for diagnostics.
+package zpages
 
 import (
 	"flag"
